refactor: mv fetcher, contexts
