--- src/components/common/fetchData.js
@@ -1,411 +1,0 @@
-import { category } from '@/constants/category'
-
-export async function getContentList(params) {
-  const {
-    topic = null,
-    page = 1,
-    tag_id = null,
-    search = ''
-  } = (params || {});
-
-  const url = new URL(`${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/content/list`);
-
-  url.searchParams.append('pageID', page);
-  const categoryID = category[topic] || "";
-  categoryID && url.searchParams.append('contents_type', categoryID);
-  tag_id && url.searchParams.append('tag_id[]', tag_id);
-  search && url.searchParams.append('filter', `keyword contains ${search}`);
-
-  const res = await fetch(url);
-  return await res.json();
-}
-
-export async function getAllContentList() {
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/content/all`
-  );
-  const data = await res.json();
-  return data.list;
-}
-
-export async function getCategoryList() {
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/category/list`
-  );
-  const data = await res.json();
-  return data.list;
-}
-
-export async function getDetails(id) {
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/content/details/${id}`,
-    {
-      method: "GET",
-      headers: { "Content-Type": "application/json" },
-      credentials: "include"
-    }
-  );
-  const data = await res.json();
-  return data.details;
-}
-
-export async function getFeatureList() {
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/pickup/list`
-  );
-  const data = await res.json();
-  return data.list;
-}
-
-export async function getFeatureDetails(id) {
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/pickup/details/${id}`
-  );
-  const data = await res.json();
-  return data.details;
-}
-
-export async function getLimitedContent() {
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/limited-content`,
-    {
-      method: 'GET',
-      headers: { 'Content-Type': 'application/json' },
-      credentials: 'include'
-    },
-  )
-  const data = await res.json()
-  return data.list
-}
-
-export async function getLimitedContentDetails(id) {
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/limited-content/details/${id}`,
-    {
-      method: 'GET',
-      headers: { 'Content-Type': 'application/json' },
-      credentials: 'include'
-    },
-  )
-  const data = await res.json()
-  return data.details
-}
-
-export async function getTagArea() {
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/tag/area`
-  );
-  const data = await res.json();
-  return data.list;
-}
-
-export async function getTagKeyword() {
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/tag/keyword`
-  );
-  const data = await res.json();
-  return data.list;
-}
-
-export async function getRanking() {
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/ranking`
-  );
-  const data = await res.json();
-  return data.list;
-}
-
-export async function getMyFavoriteList() {
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/my_favorite_list`,
-    {
-      method: 'GET',
-      // body: JSON.stringify(params),
-      headers: { "Content-Type": "application/json" },
-      credentials: "include"
-    }
-  );
-  const data = await res.json();
-  return data;
-}
-
-export async function postFavorite(module_id) {
-  const params = {
-    module_id,
-  }
-
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/favorite/register`,
-    {
-      method: 'POST',
-      body: JSON.stringify(params),
-      headers: { "Content-Type": "application/json" },
-      credentials: "include"
-    }
-  );
-
-  await res.json()
-  if (res.ok) {
-    return res
-  }
-
-  return null
-}
-
-export async function deleteFavorite(module_id) {
-  const params = {
-    module_id,
-  }
-
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/favorite/delete`,
-    {
-      method: 'POST',
-      body: JSON.stringify(params),
-      headers: { 'Content-Type': 'application/json' },
-      credentials: 'include',
-    },
-  )
-
-  await res.json()
-  if (res.ok) {
-    return res
-  }
-
-  return null
-}
-
-export async function login(email, password) {
-  const credentials = {
-    email,
-    password,
-  }
-
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/login`,
-    {
-      method: 'POST',
-      body: JSON.stringify(credentials),
-      headers: { 'Content-Type': 'application/json' },
-      credentials: 'include',
-    },
-  )
-
-  await res.json()
-
-  if (!res.ok) {
-    return null
-  }
-
-  const user = await profile()
-  if (res.ok && user) {
-    return user
-  }
-
-  return null
-}
-
-export async function getInquiryColumns() {
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/inquiry/1`,
-    {
-      method: 'GET',
-      headers: { 'Content-Type': 'application/json' },
-      credentials: 'include',
-    }
-  );
-
-  const data = await res.json()
-  return data.details
-}
-
-export async function inquiry(formData) {
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/inquiry/1`,
-    {
-      method: "POST",
-      body: JSON.stringify(formData),
-      headers: { "Content-Type": "application/json" },
-    }
-  );
-
-  const status = await res.json();
-  return status;
-}
-
-export async function uploadFile(fileData) {
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/upload`,
-    {
-      method: "POST",
-      body: fileData,
-    }
-  );
-
-  const status = await res.json();
-  return status;
-}
-
-export async function register(name1, name2, email, login_pwd) {
-  const credentials = {
-    name1,
-    name2,
-    email,
-    login_pwd,
-  }
-
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/member/register`,
-    {
-      method: 'POST',
-      body: JSON.stringify(credentials),
-      headers: { 'Content-Type': 'application/json' },
-      credentials: 'include',
-    },
-  )
-
-  await res.json()
-  if (!res.ok) {
-    return null
-  }
-
-  const user = await login(email, login_pwd)
-  if (user) {
-    return user
-  }
-  return null
-}
-
-export async function logout() {
-  await fetch(`${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/logout`, {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    credentials: 'include',
-  })
-
-  return null
-}
-
-export async function reminder(email) {
-  const credentials = {
-    email,
-  }
-
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/reminder`,
-    {
-      method: 'POST',
-      body: JSON.stringify(credentials),
-      headers: { 'Content-Type': 'application/json' },
-      credentials: 'include',
-    },
-  )
-
-  const response = await res.json();
-  if (res.ok) {
-    return response.messages;
-  }
-
-  return null
-}
-
-export async function reset(token, temp_pwd, login_pwd) {
-  const credentials = {
-    token,
-    temp_pwd,
-    login_pwd,
-  }
-
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/reminder`,
-    {
-      method: 'POST',
-      body: JSON.stringify(credentials),
-      headers: { 'Content-Type': 'application/json' },
-      credentials: 'include',
-    },
-  )
-
-  const response = await res.json();
-
-  if (res.ok) {
-    return response.messages;
-  }
-
-  return null
-}
-
-export async function profile() {
-  const userRef = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/profile`,
-    {
-      method: 'GET',
-      headers: { 'Content-Type': 'application/json' },
-      credentials: 'include',
-    },
-  )
-
-  const user = await userRef.json()
-  if (user) {
-    return user
-  }
-
-  return null
-}
-
-export async function getMemberInfo() {
-  const userRef = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/member/me`,
-    {
-      method: 'GET',
-      headers: { 'Content-Type': 'application/json' },
-      credentials: 'include',
-    },
-  )
-
-  const user = await userRef.json()
-  if (user) {
-    return user
-  }
-  return null
-}
-
-export async function updateMemberInfo(name1, name2, email, current_password, login_pwd) {
-  const credentials = {
-    name1,
-    name2,
-    email,
-    current_password,
-    login_pwd,
-  }
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/member/update`,
-    {
-      method: 'POST',
-      body: JSON.stringify(credentials),
-      headers: { 'Content-Type': 'application/json' },
-      credentials: 'include',
-    },
-  )
-
-  const status = await res.json();
-  return status;
-}
-
-export async function deleteMember() {
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/member/delete`,
-    {
-      method: 'POST',
-      body: JSON.stringify({}),
-      headers: { 'Content-Type': 'application/json' },
-      credentials: 'include',
-    },
-  )
-
-  const status = await res.json()
-  if (status.messages) {
-    return true
-  }
-  return null
-}
