+++ src/fetch/getDetails.js
@@ -1,0 +1,12 @@
+export default async function getDetails(id) {
+    const res = await fetch(
+        `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/content/details/${id}`,
+        {
+            method: "GET",
+            headers: { "Content-Type": "application/json" },
+            credentials: "include"
+        }
+    );
+    const data = await res.json();
+    return data.details;
+}
