+++ src/fetch/getInquiryColumns.js
@@ -1,0 +1,13 @@
+export default async function getInquiryColumns() {
+    const res = await fetch(
+        `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/inquiry/1`,
+        {
+            method: 'GET',
+            headers: { 'Content-Type': 'application/json' },
+            credentials: 'include',
+        }
+    );
+
+    const data = await res.json()
+    return data.details
+}
