+++ src/fetch/getMemberMe.js
@@ -1,0 +1,11 @@
+export default async function getMemberMe() {
+    const res = await fetch(
+        `${process.env.NEXT_PUBLIC_BASE_URL}/rcms-api/1/member/me`,
+        {
+            method: 'GET',
+            headers: { 'Content-Type': 'application/json' },
+            credentials: 'include',
+        },
+    )
+    return await res.json();
+}
